pyproject.toml
[project]
name = "qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents-cli>=0.0.25",
    "jsonpatch>=1.33",
    "langchain-core>=1.2.16",
    "langchain-openrouter>=0.0.2",
    "rich>=14.3.3",
]
